Code refactoring, null-safety
