+++ app/build.gradle.kts
@@ -4,14 +4,12 @@ plugins {
 
 android {
     compileSdkVersion(30)
-//    buildToolsVersion "30.0.2"
     defaultConfig {
         applicationId = "ru.commandos.diner.delivery"
         minSdkVersion(16)
         targetSdkVersion(30)
         versionCode = 1
         versionName = "1.0"
-
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
