+++ app/src/main/java/ru/commandos/diner/delivery/MainActivity.java
