+++ app/src/main/java/ru/commandos/diner/delivery/controller/CourierService.java
@@ -1,13 +1,18 @@
 package ru.commandos.diner.delivery.controller;
 
+import androidx.annotation.Nullable;
+
 import retrofit2.Retrofit;
 import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory;
 import retrofit2.converter.gson.GsonConverterFactory;
+import retrofit2.internal.EverythingIsNonNull;
 import ru.commandos.diner.delivery.BuildConfig;
 
+@EverythingIsNonNull
 public class CourierService {
 
     private static final String BASE_URL = "https://10.0.2.2:8080";
+    @Nullable
     private static CourierService courierService;
     private final Retrofit retrofit;
